[executor] Turn Executor abstraction into a concrete class. NFC.
We're going to rewrite `Executor` heavily. Each field execution will
turn into a `concurrent.Task` and will be executed by
`concurrent.Executor`. `Executor` will become complicated and more
self-contained. So we decide to remove `Executor` abstraction.
